feat: add description to VIP-224
